requirements.txt
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==3.0.4
discord==1.0.1
discord.py==1.6.0
idna==2.10
multidict==5.1.0
PyYAML==5.4.1
requests==2.25.1
soupsieve==2.2
typing-extensions==3.7.4.3
urllib3==1.26.3
yarl==1.6.3
